move keycodes.h to qcommon
